feat(backend): enhance error handling and add logging
- Introduced custom APIException class for structured error responses.
- Implemented global and API-specific exception handlers in main.py.
- Added loguru for logging errors.
- Removed outdated core exceptions and refactored job description service to use new exception handling.
- Added constants for job description prompts.

Refs: #8
