+++ backend/exceptions.py
@@ -1,0 +1,5 @@
+from fastapi import HTTPException
+
+
+class APIException(HTTPException):
+    """Base exception class for all API exceptions."""
