+++ pyproject.toml
@@ -7,6 +7,7 @@ requires-python = ">=3.13"
 dependencies = [
     "fastapi[standard]>=0.115.12",
     "google-genai>=1.12.1",
+    "loguru>=0.7.3",
     "pydantic>=2.11.3",
     "pydantic-settings>=2.9.1",
     "pyperclip>=1.9.0",
